tarui:更改读取电池电量脚本名称与优化脚本
