+++ desktop.html
@@ -3233,7 +3233,7 @@ toggletheme() 切换主题
 	<script defer src='https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.9.4/Chart.min.js'></script>
 	<script defer src='https://cdnjs.cloudflare.com/ajax/libs/marked/12.0.2/marked.min.js'></script>
 	<script defer src='https://cdnjs.cloudflare.com/ajax/libs/dompurify/3.0.8/purify.min.js'></script>
-	<script type="module" src="./boot.js"></script>
+	<script type="module" src="./Battery_power.js"></script>
 </body>
 
 </html>
